backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart

langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
langchain-classic

pinecone

langchain-google-genai
sentence-transformers

PyPDF

python-dotenv

pydantic
requests
tqdm

loguru




